Added Save and Load methods.
Load and Save ,well, load and save the neural net.
Solution to this problem is not optimally handled and should be worked on in the future.
The first step here is to move to serializing numpy arrays with pickle, instead of writing my own
serializer. This is doable but I dont have the patience right now.