Add .gitignore. Modify files to enable upload to pypi.
Added a standard python gitignore file.

Fixed the setup.py script. The problem was that
version of dendrites was kept in the dendrites
module (__version__).

However, to install dendrites, setup.py requires info about version,
and it was importing dendrites and reading the __version__variable.

This was a serious issue since setup.py is importing dendrites
before it installed all the dependencies, thus leading to failed
imports.

(Specifically, setup.py was importing dendrites, dendrites were
importing numpy, and numpy wasn't installed. Setup py is supposed to install it
later.)